refactor: Refactor the legend details.
