+++ packages/map-template/src/components/LegendDialog/LegendDialog.jsx
@@ -9,6 +9,7 @@ import { useIsKioskContext } from "../../hooks/useIsKioskContext";
 import { createPortal } from "react-dom";
 import legendSizeState from "../../atoms/legendSizeState";
 import getLegendSectionsHeight from "../../helpers/GetLegendSectionsHeight";
+import getLegendSortedFields from "../../helpers/GetLegendSortedFields";
 
 /**
  * Handle the Legend dialog.
@@ -20,7 +21,7 @@ function LegendDialog() {
 
     const kioskLocation = useRecoilValue(kioskLocationState)
 
-    const [legendSections, setLegendSections] = useState([]);
+    const legendSections = useState(getLegendSortedFields(kioskLocation));
 
     const isKioskContext = useIsKioskContext();
 
@@ -35,42 +36,6 @@ function LegendDialog() {
     const [legendSize, setLegendSize] = useRecoilState(legendSizeState);
 
     const [showScrollButtons, setShowScrollButtons] = useState(false);
-
-    useEffect(() => {
-        if (!kioskLocation?.properties.fields) return;
-
-        const legendFields = [];
-
-        for (const customPropertyKey of Object.keys(kioskLocation.properties.fields)) {
-            const index = parseInt(customPropertyKey.charAt(0));
-            // Skip field if first character isn't a number.
-            if (!Number.isInteger(index)) continue;
-
-            const existingEntry = legendFields.find(i => i.index === index);
-            const isHeadingEntry = customPropertyKey.toLowerCase().includes('legendheading');
-            const isContentEntry = customPropertyKey.toLowerCase().includes('legendcontent');
-
-            if (!existingEntry) {
-                // Create new legend field
-                const newEntry = {
-                    index: index,
-                    heading: isHeadingEntry ? kioskLocation.properties.fields[customPropertyKey].value : null,
-                    content: isContentEntry ? kioskLocation.properties.fields[customPropertyKey].value : null
-                };
-                legendFields.push(newEntry);
-            } else {
-                // Update existing legend field
-                if (isHeadingEntry) {
-                    existingEntry.heading = kioskLocation.properties.fields[customPropertyKey].value;
-                } else if (isContentEntry) {
-                    existingEntry.content = kioskLocation.properties.fields[customPropertyKey].value;
-                }
-            }
-        }
-
-        const sortedFields = legendFields.sort((a, b) => a.index - b.index);
-        setLegendSections(sortedFields);
-    }, [kioskLocation]);
 
     /*
      * Get the height of the legend sections. 
@@ -87,7 +52,6 @@ function LegendDialog() {
                     setShowScrollButtons(false);
                 }
             });
-
         }
     }, [showLegendDialog, isKioskContext]);
